Use unsigned int for partition blocks free/available counts.
Avoid useless overflow when scaling from statvfs results by going
straight to 1K blocks instead of using a 512-byte block intermediate
step. This should allow us to use up to 4TB partitions instead of
the current 1TB limit.

+++ coda-src/volutil/backup.cc
@@ -552,7 +552,7 @@ struct DiskPartition *findBestPartition(void)
 	tmp = &DiskPartitionList;
 	while( (tmp = tmp->next) != &DiskPartitionList) {
 		part = list_entry(tmp, struct DiskPartition, dp_chain);
-		if (!best || part->free > (long)space) {
+		if (!best || part->free > space) {
 			best = part;
 			space = part->free;
 		}
